Through meta tests for #16.
