+++ afscgap/model.py
@@ -1,164 +1,44 @@
 """
 Logic for representing and working with data returned by the AFSC GAP API.
 
-(c) 2023 The Eric and Wendy Schmidt Center for Data Science and the Environment
-at UC Berkeley.
+(c) 2023 Regents of University of California / The Eric and Wendy Schmidt Center
+for Data Science and the Environment at UC Berkeley.
 
 This file is part of afscgap released under the BSD 3-Clause License. See
 LICENSE.txt.
 """
-import re
-import typing
-
 from afscgap.util import OPT_FLOAT
 from afscgap.util import OPT_INT
 
-DATE_REGEX = re.compile('(?P<month>\\d{2})\\/(?P<day>\\d{2})\\/' + \
-    '(?P<year>\\d{4}) (?P<hours>\\d{2})\\:(?P<minutes>\\d{2})\\:' + \
-    '(?P<seconds>\\d{2})')
-DATE_TEMPLATE = '%s/%s/%s %s:%s:%s'
-ISO_8601_REGEX = re.compile('(?P<year>\\d{4})\\-(?P<month>\\d{2})\\-' + \
-    '(?P<day>\\d{2})T(?P<hours>\\d{2})\\:(?P<minutes>\\d{2})\\:' + \
-    '(?P<seconds>\\d{2})')
-ISO_8601_TEMPLATE = '%s-%s-%sT%s:%s:%s'
+OPT_RECORD = 'typing.Optional[Record]'
 
 
 class HaulKeyable:
-    
+
     def get_srvy(self) -> str:
         raise NotImplementedError('Use implementor.')
-    
+
     def get_year(self) -> float:
         raise NotImplementedError('Use implementor.')
-    
+
     def get_vessel_id(self) -> float:
         raise NotImplementedError('Use implementor.')
-    
+
     def get_cruise(self) -> float:
         raise NotImplementedError('Use implementor.')
-    
+
     def get_haul(self) -> float:
         raise NotImplementedError('Use implementor.')
 
 
 class Record(HaulKeyable):
-    """Data structure describing a single record returned by the API."""
+    """Interface describing a single record.
 
-    def __init__(self, year: float, srvy: str, survey: str, survey_id: float,
-        cruise: float, haul: float, stratum: float, station: str,
-        vessel_name: str, vessel_id: float, date_time: str, latitude_dd: float,
-        longitude_dd: float, species_code: float, common_name: str,
-        scientific_name: str, taxon_confidence: str, cpue_kgha: OPT_FLOAT,
-        cpue_kgkm2: OPT_FLOAT, cpue_kg1000km2: OPT_FLOAT, cpue_noha: OPT_FLOAT,
-        cpue_nokm2: OPT_FLOAT, cpue_no1000km2: OPT_FLOAT, weight_kg: OPT_FLOAT,
-        count: OPT_FLOAT, bottom_temperature_c: OPT_FLOAT,
-        surface_temperature_c: OPT_FLOAT, depth_m: float,
-        distance_fished_km: float, net_width_m: float, net_height_m: float,
-        area_swept_ha: float, duration_hr: float, tsn: OPT_INT,
-        ak_survey_id: int):
-        """Create a new record.
-
-        Args:
-            year: Year for the survey in which this observation was made.
-            srvy: The name of the survey in which this observation was made.
-                NBS (N Bearing Sea), EBS (SE Bearing Sea), BSS (Bearing Sea
-                Slope), or GOA (Gulf of Alaska)
-            survey: Long form description of the survey in which the observation
-                was made.
-            survey_id: Unique numeric ID for the survey.
-            cruise: An ID uniquely identifying the cruise in which the
-                observation was made. Multiple cruises in a survey.
-            haul: An ID uniquely identifying the haul in which this observation
-                was made. Multiple hauls per cruises.
-            stratum: Unique ID for statistical area / survey combination as
-                described in the metadata or 0 if an experimental tow.
-            station: Station associated with the survey.
-            vessel_name: Unique ID describing the vessel that made this
-                observation. This is left as a string but, in practice, is
-                likely numeric.
-            vessel_id: Name of the vessel at the time the observation was made
-                with multiple names potentially associated with a vessel ID.
-            date_time: The date and time of the haul which has been attempted to
-                be transformed to an ISO 8601 string without timezone info.
-                If it couldn’t be transformed, the original string.
-            latitude_dd: Latitude in decimal degrees associated with the haul.
-            longitude_dd: Longitude in decimal degrees associated with the haul.
-            species_code: Unique ID associated with the species observed.
-            common_name: The “common name” associated with the species observed.
-            scientific_name: The “scientific name” associated with the species
-                observed.
-            taxon_confidence: Confidence flag regarding ability to identify
-                species (High, Moderate, Low). In practice, this can also be
-                Unassessed.
-            cpue_kgha: Catch weight divided by net area (kg / hectares) if
-                available. None if could not interpret as a float.
-            cpue_kgkm2: Catch weight divided by net area (kg / km^2) if
-                available. None if could not interpret as a float.
-            cpue_kg1000km2: Catch weight divided by net area (kg / km^2 * 1000)
-                if available. See metadata. None if could not interpret as a
-                float.
-            cpue_noha: Catch number divided by net sweep area if available
-                (count / hectares). None if could not interpret as a float.
-            cpue_nokm2: Catch number divided by net sweep area if available
-                (count / km^2). See metadata. None if could not interpret as a
-                float.
-            cpue_no1000km2: Catch number divided by net sweep area if available
-                (count / km^2 * 1000). See metadata. None if could not interpret
-                as a float.
-            weight_kg: Taxon weight (kg) if available. See metadata. None if
-                could not interpret as a float.
-            count: Total number of organism individuals in haul. None if could
-                not interpret as a float.
-            bottom_temperature_c: Bottom temperature associated with observation
-                if available in Celsius. None if not given or could not
-                interpret as a float.
-            surface_temperature_c: Surface temperature associated with
-                observation if available in Celsius. None if not given or could
-                not interpret as a float.
-            depth_m: Depth of the bottom in meters.
-            distance_fished_km: Distance of the net fished as km.
-            net_width_m: Distance of the net fished as m.
-            net_height_m: Height of the net fished as m.
-            area_swept_ha: Area covered by the net while fishing in hectares.
-            duration_hr: Duration of the haul as number of hours.
-            tsn: Taxonomic information system species code.
-            ak_survey_id: AK identifier for the survey.
-        """
-        self._year = year
-        self._srvy = srvy
-        self._survey = survey
-        self._survey_id = survey_id
-        self._cruise = cruise
-        self._haul = haul
-        self._stratum = stratum
-        self._station = station
-        self._vessel_name = vessel_name
-        self._vessel_id = vessel_id
-        self._date_time = date_time
-        self._latitude_dd = latitude_dd
-        self._longitude_dd = longitude_dd
-        self._species_code = species_code
-        self._common_name = common_name
-        self._scientific_name = scientific_name
-        self._taxon_confidence = taxon_confidence
-        self._cpue_kgha = cpue_kgha
-        self._cpue_kgkm2 = cpue_kgkm2
-        self._cpue_kg1000km2 = cpue_kg1000km2
-        self._cpue_noha = cpue_noha
-        self._cpue_nokm2 = cpue_nokm2
-        self._cpue_no1000km2 = cpue_no1000km2
-        self._weight_kg = weight_kg
-        self._count = count
-        self._bottom_temperature_c = bottom_temperature_c
-        self._surface_temperature_c = surface_temperature_c
-        self._depth_m = depth_m
-        self._distance_fished_km = distance_fished_km
-        self._net_width_m = net_width_m
-        self._net_height_m = net_height_m
-        self._area_swept_ha = area_swept_ha
-        self._duration_hr = duration_hr
-        self._tsn = tsn
-        self._ak_survey_id = ak_survey_id
+    Interface describing a single record of an observtion. Note that, in
+    practice, this "observation" can be a presence obervation where a species
+    was found or an "absence" / "zero catch" observation where a sepcies was
+    not observed in a haul.
+    """
 
     def get_year(self) -> float:
         """Get the field labeled as year in the API.
@@ -166,7 +46,7 @@ class Record(HaulKeyable):
         Returns:
             Year for the survey in which this observation was made.
         """
-        return self._year
+        raise NotImplementedError('Use implementor.')
 
     def get_srvy(self) -> str:
         """Get the field labeled as srvy in the API.
@@ -176,7 +56,7 @@ class Record(HaulKeyable):
             Bearing Sea), EBS (SE Bearing Sea), BSS (Bearing Sea Slope), or GOA
             (Gulf of Alaska)
         """
-        return self._srvy
+        raise NotImplementedError('Use implementor.')
 
     def get_survey(self) -> str:
         """Get the field labeled as survey in the API.
@@ -185,7 +65,7 @@ class Record(HaulKeyable):
             Long form description of the survey in which the observation was
             made.
         """
-        return self._survey
+        raise NotImplementedError('Use implementor.')
 
     def get_survey_id(self) -> float:
         """Get the field labeled as survey_id in the API.
@@ -193,7 +73,7 @@ class Record(HaulKeyable):
         Returns:
             Unique numeric ID for the survey.
         """
-        return self._survey_id
+        raise NotImplementedError('Use implementor.')
 
     def get_cruise(self) -> float:
         """Get the field labeled as cruise in the API.
@@ -202,7 +82,7 @@ class Record(HaulKeyable):
             An ID uniquely identifying the cruise in which the observation was
             made. Multiple cruises in a survey.
         """
-        return self._cruise
+        raise NotImplementedError('Use implementor.')
 
     def get_haul(self) -> float:
         """Get the field labeled as haul in the API.
@@ -211,7 +91,7 @@ class Record(HaulKeyable):
             An ID uniquely identifying the haul in which this observation was
             made. Multiple hauls per cruises.
         """
-        return self._haul
+        raise NotImplementedError('Use implementor.')
 
     def get_stratum(self) -> float:
         """Get the field labeled as stratum in the API.
@@ -220,7 +100,7 @@ class Record(HaulKeyable):
             Unique ID for statistical area / survey combination as described in
             the metadata or 0 if an experimental tow.
         """
-        return self._stratum
+        raise NotImplementedError('Use implementor.')
 
     def get_station(self) -> str:
         """Get the field labeled as station in the API.
@@ -228,7 +108,7 @@ class Record(HaulKeyable):
         Returns:
             Station associated with the survey.
         """
-        return self._station
+        raise NotImplementedError('Use implementor.')
 
     def get_vessel_name(self) -> str:
         """Get the field labeled as vessel_name in the API.
@@ -237,7 +117,7 @@ class Record(HaulKeyable):
             Unique ID describing the vessel that made this observation. This is
             left as a string but, in practice, is likely numeric.
         """
-        return self._vessel_name
+        raise NotImplementedError('Use implementor.')
 
     def get_vessel_id(self) -> float:
         """Get the field labeled as vessel_id in the API.
@@ -246,7 +126,7 @@ class Record(HaulKeyable):
             Name of the vessel at the time the observation was made with
             multiple names potentially associated with a vessel ID.
         """
-        return self._vessel_id
+        raise NotImplementedError('Use implementor.')
 
     def get_date_time(self) -> str:
         """Get the field labeled as date_time in the API.
@@ -256,7 +136,7 @@ class Record(HaulKeyable):
             transformed to an ISO 8601 string without timezone info. If it
             couldn’t be transformed, the original string is reported.
         """
-        return self._date_time
+        raise NotImplementedError('Use implementor.')
 
     def get_latitude_dd(self) -> float:
         """Get the field labeled as latitude_dd in the API.
@@ -264,7 +144,7 @@ class Record(HaulKeyable):
         Returns:
             Latitude in decimal degrees associated with the haul.
         """
-        return self._latitude_dd
+        raise NotImplementedError('Use implementor.')
 
     def get_longitude_dd(self) -> float:
         """Get the field labeled as longitude_dd in the API.
@@ -272,7 +152,7 @@ class Record(HaulKeyable):
         Returns:
             Longitude in decimal degrees associated with the haul.
         """
-        return self._longitude_dd
+        raise NotImplementedError('Use implementor.')
 
     def get_species_code(self) -> float:
         """Get the field labeled as species_code in the API.
@@ -280,7 +160,7 @@ class Record(HaulKeyable):
         Returns:
             Unique ID associated with the species observed.
         """
-        return self._species_code
+        raise NotImplementedError('Use implementor.')
 
     def get_common_name(self) -> str:
         """Get the field labeled as common_name in the API.
@@ -289,7 +169,7 @@ class Record(HaulKeyable):
             The “common name” associated with the species observed. Example:
             Pacific glass shrimp.
         """
-        return self._common_name
+        raise NotImplementedError('Use implementor.')
 
     def get_scientific_name(self) -> str:
         """Get the field labeled as scientific_name in the API.
@@ -298,7 +178,7 @@ class Record(HaulKeyable):
             The “scientific name” associated with the species observed. Example:
             Pasiphaea pacifica.
         """
-        return self._scientific_name
+        raise NotImplementedError('Use implementor.')
 
     def get_taxon_confidence(self) -> str:
         """Get the field labeled as taxon_confidence in the API.
@@ -307,7 +187,7 @@ class Record(HaulKeyable):
             Confidence flag regarding ability to identify species (High,
             Moderate, Low). In practice, this can also be Unassessed.
         """
-        return self._taxon_confidence
+        raise NotImplementedError('Use implementor.')
 
     def get_cpue_kgha_maybe(self) -> OPT_FLOAT:
         """Get the field labeled as cpue_kgha in the API.
@@ -316,7 +196,7 @@ class Record(HaulKeyable):
             Catch weight divided by net area (kg / hectares) if available. See
             metadata. None if could not interpret as a float.
         """
-        return self._cpue_kgha
+        raise NotImplementedError('Use implementor.')
 
     def get_cpue_kgkm2_maybe(self) -> OPT_FLOAT:
         """Get the field labeled as cpue_kgkm2 in the API.
@@ -325,7 +205,7 @@ class Record(HaulKeyable):
             Catch weight divided by net area (kg / km^2) if available. See
             metadata. None if could not interpret as a float.
         """
-        return self._cpue_kgkm2
+        raise NotImplementedError('Use implementor.')
 
     def get_cpue_kg1000km2_maybe(self) -> OPT_FLOAT:
         """Get the field labeled as cpue_kg1000km2 in the API.
@@ -334,7 +214,7 @@ class Record(HaulKeyable):
             Catch weight divided by net area (kg / km^2 * 1000) if available.
             See metadata. None if could not interpret as a float.
         """
-        return self._cpue_kg1000km2
+        raise NotImplementedError('Use implementor.')
 
     def get_cpue_noha_maybe(self) -> OPT_FLOAT:
         """Get the field labeled as cpue_noha in the API.
@@ -343,7 +223,7 @@ class Record(HaulKeyable):
             Catch number divided by net sweep area if available (count /
             hectares). See metadata. None if could not interpret as a float.
         """
-        return self._cpue_noha
+        raise NotImplementedError('Use implementor.')
 
     def get_cpue_nokm2_maybe(self) -> OPT_FLOAT:
         """Get the field labeled as cpue_nokm2 in the API.
@@ -352,7 +232,7 @@ class Record(HaulKeyable):
             Catch number divided by net sweep area if available (count / km^2).
             See metadata. None if could not interpret as a float.
         """
-        return self._cpue_nokm2
+        raise NotImplementedError('Use implementor.')
 
     def get_cpue_no1000km2_maybe(self) -> OPT_FLOAT:
         """Get the field labeled as cpue_no1000km2 in the API.
@@ -361,7 +241,7 @@ class Record(HaulKeyable):
             Catch number divided by net sweep area if available (count / km^2 *
             1000). See metadata. None if could not interpret as a float.
         """
-        return self._cpue_no1000km2
+        raise NotImplementedError('Use implementor.')
 
     def get_weight_kg_maybe(self) -> OPT_FLOAT:
         """Get the field labeled as weight_kg in the API.
@@ -370,7 +250,7 @@ class Record(HaulKeyable):
             Taxon weight (kg) if available. See metadata. None if could not
             interpret as a float.
         """
-        return self._weight_kg
+        raise NotImplementedError('Use implementor.')
 
     def get_count_maybe(self) -> OPT_FLOAT:
         """Get the field labeled as count in the API.
@@ -379,7 +259,7 @@ class Record(HaulKeyable):
             Total number of organism individuals in haul. None if could not
             interpret as a float.
         """
-        return self._count
+        raise NotImplementedError('Use implementor.')
 
     def get_bottom_temperature_c_maybe(self) -> OPT_FLOAT:
         """Get the field labeled as bottom_temperature_c in the API.
@@ -388,7 +268,7 @@ class Record(HaulKeyable):
             Bottom temperature associated with observation if available in
             Celsius. None if not given or could not interpret as a float.
         """
-        return self._bottom_temperature_c
+        raise NotImplementedError('Use implementor.')
 
     def get_surface_temperature_c_maybe(self) -> OPT_FLOAT:
         """Get the field labeled as surface_temperature_c in the API.
@@ -397,7 +277,7 @@ class Record(HaulKeyable):
             Surface temperature associated with observation if available in
             Celsius. None if not given or could not interpret as a float.
         """
-        return self._surface_temperature_c
+        raise NotImplementedError('Use implementor.')
 
     def get_depth_m(self) -> float:
         """Get the field labeled as depth_m in the API.
@@ -405,7 +285,7 @@ class Record(HaulKeyable):
         Returns:
             Depth of the bottom in meters.
         """
-        return self._depth_m
+        raise NotImplementedError('Use implementor.')
 
     def get_distance_fished_km(self) -> float:
         """Get the field labeled as distance_fished_km in the API.
@@ -413,7 +293,7 @@ class Record(HaulKeyable):
         Returns:
             Distance of the net fished as km.
         """
-        return self._distance_fished_km
+        raise NotImplementedError('Use implementor.')
 
     def get_net_width_m(self) -> float:
         """Get the field labeled as net_width_m in the API.
@@ -421,7 +301,7 @@ class Record(HaulKeyable):
         Returns:
             Distance of the net fished as m.
         """
-        return self._net_width_m
+        raise NotImplementedError('Use implementor.')
 
     def get_net_height_m(self) -> float:
         """Get the field labeled as net_height_m in the API.
@@ -429,7 +309,7 @@ class Record(HaulKeyable):
         Returns:
             Height of the net fished as m.
         """
-        return self._net_height_m
+        raise NotImplementedError('Use implementor.')
 
     def get_area_swept_ha(self) -> float:
         """Get the field labeled as area_swept_ha in the API.
@@ -437,7 +317,7 @@ class Record(HaulKeyable):
         Returns:
             Area covered by the net while fishing in hectares.
         """
-        return self._area_swept_ha
+        raise NotImplementedError('Use implementor.')
 
     def get_duration_hr(self) -> float:
         """Get the field labeled as duration_hr in the API.
@@ -445,7 +325,7 @@ class Record(HaulKeyable):
         Returns:
             Duration of the haul as number of hours.
         """
-        return self._duration_hr
+        raise NotImplementedError('Use implementor.')
 
     def get_tsn(self) -> int:
         """Get the field labeled as tsn in the API.
@@ -453,7 +333,7 @@ class Record(HaulKeyable):
         Returns:
             Taxonomic information system species code.
         """
-        return assert_int_present(self._tsn)
+        raise NotImplementedError('Use implementor.')
 
     def get_tsn_maybe(self) -> OPT_INT:
         """Get the field labeled as tsn in the API or None.
@@ -462,7 +342,7 @@ class Record(HaulKeyable):
             Taxonomic information system species code if it could be parsed as
             an int and None otherwise.
         """
-        return self._tsn
+        raise NotImplementedError('Use implementor.')
 
     def get_ak_survey_id(self) -> int:
         """Get the field labeled as ak_survey_id in the API.
@@ -470,7 +350,7 @@ class Record(HaulKeyable):
         Returns:
             AK identifier for the survey.
         """
-        return self._ak_survey_id
+        raise NotImplementedError('Use implementor.')
 
     def get_ak_survey_id_maybe(self) -> OPT_INT:
         """Get the field labeled as ak_survey_id in the API.
@@ -478,7 +358,7 @@ class Record(HaulKeyable):
         Returns:
             AK identifier for the survey or None if not given.
         """
-        return self._ak_survey_id
+        raise NotImplementedError('Use implementor.')
 
     def get_cpue_kgha(self) -> float:
         """Get the value of field cpue_kgha with validity assert.
@@ -491,7 +371,7 @@ class Record(HaulKeyable):
             Catch weight divided by net area (kg / hectares) if available. See
             metadata.
         """
-        return assert_float_present(self._cpue_kgha)
+        raise NotImplementedError('Use implementor.')
 
     def get_cpue_kgkm2(self) -> float:
         """Get the value of field cpue_kgkm2 with validity assert.
@@ -504,7 +384,7 @@ class Record(HaulKeyable):
             Catch weight divided by net area (kg / km^2) if available. See
             metadata.
         """
-        return assert_float_present(self._cpue_kgkm2)
+        raise NotImplementedError('Use implementor.')
 
     def get_cpue_kg1000km2(self) -> float:
         """Get the value of field cpue_kg1000km2 with validity assert.
@@ -517,7 +397,7 @@ class Record(HaulKeyable):
             Catch weight divided by net area (kg / km^2 * 1000) if available.
             See metadata.
         """
-        return assert_float_present(self._cpue_kg1000km2)
+        raise NotImplementedError('Use implementor.')
 
     def get_cpue_noha(self) -> float:
         """Get the value of field cpue_noha with validity assert.
@@ -530,7 +410,7 @@ class Record(HaulKeyable):
             Catch number divided by net sweep area if available (count /
             hectares). See metadata.
         """
-        return assert_float_present(self._cpue_noha)
+        raise NotImplementedError('Use implementor.')
 
     def get_cpue_nokm2(self) -> float:
         """Get the value of field cpue_nokm2 with validity assert.
@@ -543,7 +423,7 @@ class Record(HaulKeyable):
             Catch number divided by net sweep area if available (count / km^2).
             See metadata.
         """
-        return assert_float_present(self._cpue_nokm2)
+        raise NotImplementedError('Use implementor.')
 
     def get_cpue_no1000km2(self) -> float:
         """Get the value of field cpue_no1000km2 with validity assert.
@@ -556,7 +436,7 @@ class Record(HaulKeyable):
             Catch number divided by net sweep area if available (count / km^2 *
             1000). See metadata.
         """
-        return assert_float_present(self._cpue_no1000km2)
+        raise NotImplementedError('Use implementor.')
 
     def get_weight_kg(self) -> float:
         """Get the value of field weight_kg with validity assert.
@@ -568,7 +448,7 @@ class Record(HaulKeyable):
         Returns:
             Taxon weight (kg) if available. See metadata.
         """
-        return assert_float_present(self._weight_kg)
+        raise NotImplementedError('Use implementor.')
 
     def get_count(self) -> float:
         """Get the value of field count with validity assert.
@@ -580,7 +460,7 @@ class Record(HaulKeyable):
         Returns:
             Total number of organism individuals in haul.
         """
-        return assert_float_present(self._count)
+        raise NotImplementedError('Use implementor.')
 
     def get_bottom_temperature_c(self) -> float:
         """Get the value of field bottom_temperature_c with validity assert.
@@ -593,7 +473,7 @@ class Record(HaulKeyable):
             Bottom temperature associated with observation if available in
             Celsius.
         """
-        return assert_float_present(self._bottom_temperature_c)
+        raise NotImplementedError('Use implementor.')
 
     def get_surface_temperature_c(self) -> float:
         """Get the value of field surface_temperature_c with validity assert.
@@ -606,7 +486,7 @@ class Record(HaulKeyable):
             Surface temperature associated with observation if available in
             Celsius. None if not
         """
-        return assert_float_present(self._surface_temperature_c)
+        raise NotImplementedError('Use implementor.')
 
     def is_complete(self) -> bool:
         """Determine if this record has all of its values filled in.
@@ -615,25 +495,7 @@ class Record(HaulKeyable):
             True if all optional fields have a parsed value with the expected
             type and false otherwise.
         """
-        optional_fields = [
-            self._cpue_kgha,
-            self._cpue_kgkm2,
-            self._cpue_kg1000km2,
-            self._cpue_noha,
-            self._cpue_nokm2,
-            self._cpue_no1000km2,
-            self._weight_kg,
-            self._count,
-            self._bottom_temperature_c,
-            self._surface_temperature_c,
-            self._tsn
-        ]
-
-        has_none = None in optional_fields
-        all_fields_present = not has_none
-        has_valid_date_time = ISO_8601_REGEX.match(self._date_time) is not None
-
-        return all_fields_present and has_valid_date_time
+        raise NotImplementedError('Use implementor.')
 
     def to_dict(self) -> dict:
         """Serialize this Record to a dictionary form.
@@ -642,43 +504,7 @@ class Record(HaulKeyable):
             Dictionary with field names matching those found in the API results
             with incomplete records having some values as None.
         """
-        return {
-            'year': self._year,
-            'srvy': self._srvy,
-            'survey': self._survey,
-            'survey_id': self._survey_id,
-            'cruise': self._cruise,
-            'haul': self._haul,
-            'stratum': self._stratum,
-            'station': self._station,
-            'vessel_name': self._vessel_name,
-            'vessel_id': self._vessel_id,
-            'date_time': self._date_time,
-            'latitude_dd': self._latitude_dd,
-            'longitude_dd': self._longitude_dd,
-            'species_code': self._species_code,
-            'common_name': self._common_name,
-            'scientific_name': self._scientific_name,
-            'taxon_confidence': self._taxon_confidence,
-            'cpue_kgha': self._cpue_kgha,
-            'cpue_kgkm2': self._cpue_kgkm2,
-            'cpue_kg1000km2': self._cpue_kg1000km2,
-            'cpue_noha': self._cpue_noha,
-            'cpue_nokm2': self._cpue_nokm2,
-            'cpue_no1000km2': self._cpue_no1000km2,
-            'weight_kg': self._weight_kg,
-            'count': self._count,
-            'bottom_temperature_c': self._bottom_temperature_c,
-            'surface_temperature_c': self._surface_temperature_c,
-            'depth_m': self._depth_m,
-            'distance_fished_km': self._distance_fished_km,
-            'net_width_m': self._net_width_m,
-            'net_height_m': self._net_height_m,
-            'area_swept_ha': self._area_swept_ha,
-            'duration_hr': self._duration_hr,
-            'tsn': self._tsn,
-            'ak_survey_id': self._ak_survey_id,
-        }
+        raise NotImplementedError('Use implementor.')
 
 
 class Haul(HaulKeyable):
@@ -687,7 +513,7 @@ class Haul(HaulKeyable):
         haul: float, stratum: float, station: str, vessel_name: str,
         vessel_id: float, date_time: str, latitude_dd: float,
         longitude_dd: float, bottom_temperature_c: OPT_FLOAT,
-        surface_temperature_c: OPT_FLOAT, depth_m: float, 
+        surface_temperature_c: OPT_FLOAT, depth_m: float,
         distance_fished_km: float, net_width_m: float, net_height_m: float,
         area_swept_ha: float, duration_hr: float):
         self._srvy = srvy
@@ -719,17 +545,7 @@ class Haul(HaulKeyable):
         """
         return float(self.get_date_time().split('-')[0])
 
-    def get_date_time(self) -> str:
-        """Get the field labeled as date_time in the API.
-
-        Returns:
-            The date and time of the haul which has been attempted to be
-            transformed to an ISO 8601 string without timezone info. If it
-            couldn’t be transformed, the original string is reported.
-        """
-        return self._date_time
-
-   def get_srvy(self) -> str:
+    def get_srvy(self) -> str:
         """Get the field labeled as srvy in the API.
 
         Returns:
@@ -922,11 +738,35 @@ class Haul(HaulKeyable):
             Surface temperature associated with haul if available in Celsius.
         """
         return assert_float_present(self._surface_temperature_c)
-    
+
     def is_complete(self) -> bool:
         bottom_valid = self._bottom_temperature_c is not None
         surface_valid = self._surface_temperature_c is not None
         return bottom_valid and surface_valid
+
+    def to_dict(self) -> dict:
+        return {
+            'srvy': self._srvy,
+            'survey': self._survey,
+            'survey_id': self._survey_id,
+            'cruise': self._cruise,
+            'haul': self._haul,
+            'stratum': self._stratum,
+            'station': self._station,
+            'vessel_name': self._vessel_name,
+            'vessel_id': self._vessel_id,
+            'date_time': self._date_time,
+            'latitude_dd': self._latitude_dd,
+            'longitude_dd': self._longitude_dd,
+            'bottom_temperature_c': self._bottom_temperature_c,
+            'surface_temperature_c': self._surface_temperature_c,
+            'depth_m': self._depth_m,
+            'distance_fished_km': self._distance_fished_km,
+            'net_width_m': self._net_width_m,
+            'net_height_m': self._net_height_m,
+            'area_swept_ha': self._area_swept_ha,
+            'duration_hr': self._duration_hr
+        }
 
 
 def get_opt_float(target) -> OPT_FLOAT:
@@ -967,222 +807,11 @@ def get_opt_int(target) -> OPT_INT:
         return None
 
 
-def assert_float_present(self, target: OPT_FLOAT) -> float:
+def assert_float_present(target: OPT_FLOAT) -> float:
     assert target is not None
     return target
 
 
-def assert_int_present(self, target: OPT_INT) -> int:
+def assert_int_present(target: OPT_INT) -> int:
     assert target is not None
     return target
-
-
-def convert_from_iso8601(target: str) -> str:
-    """Attempt converting an ISO 8601 string to an API-provided datetime.
-
-    Args:
-        target: The datetime string to try to interpret.
-    Returns:
-        The datetime input string as a ISO 8601 string or the original value of
-        target if it could not be parsed.
-    """
-    match = ISO_8601_REGEX.match(target)
-
-    if not match:
-        return target
-
-    year = match.group('year')
-    month = match.group('month')
-    day = match.group('day')
-    hours = match.group('hours')
-    minutes = match.group('minutes')
-    seconds = match.group('seconds')
-
-    return DATE_TEMPLATE % (month, day, year, hours, minutes, seconds)
-
-
-def convert_to_iso8601(target: str) -> str:
-    """Attempt converting an API-provided datetime to ISO 8601.
-
-    Args:
-        target: The datetime string to try to interpret.
-    Returns:
-        The datetime input string as a ISO 8601 string or the original value of
-        target if it could not be parsed.
-    """
-    match = DATE_REGEX.match(target)
-
-    if not match:
-        return target
-
-    year = match.group('year')
-    month = match.group('month')
-    day = match.group('day')
-    hours = match.group('hours')
-    minutes = match.group('minutes')
-    seconds = match.group('seconds')
-
-    return ISO_8601_TEMPLATE % (year, month, day, hours, minutes, seconds)
-
-
-def parse_record(target: dict) -> Record:
-    """Parse a record from a returned item dictionary.
-
-    Args:
-        target: The dictionary from which values should be read.
-
-    Raises:
-        ValueError: Exception raised if a field has an unexpected type or cannot
-            be parsed to an expected type.
-        KeyError: Exception raised if an expected field is not found.
-
-    Returns:
-        Newly parsed record.
-    """
-    year = float(target['year'])
-    srvy = str(target['srvy'])
-    survey = str(target['survey'])
-    survey_id = float(target['survey_id'])
-    cruise = float(target['cruise'])
-    haul = float(target['haul'])
-    stratum = float(target['stratum'])
-    station = str(target['station'])
-    vessel_name = str(target['vessel_name'])
-    vessel_id = float(target['vessel_id'])
-    date_time = convert_to_iso8601(str(target['date_time']))
-    latitude_dd = float(target['latitude_dd'])
-    longitude_dd = float(target['longitude_dd'])
-    species_code = float(target['species_code'])
-    common_name = str(target['common_name'])
-    scientific_name = str(target['scientific_name'])
-    taxon_confidence = str(target['taxon_confidence'])
-    cpue_kgha = get_opt_float(target['cpue_kgha'])
-    cpue_kgkm2 = get_opt_float(target['cpue_kgkm2'])
-    cpue_kg1000km2 = get_opt_float(target['cpue_kg1000km2'])
-    cpue_noha = get_opt_float(target['cpue_noha'])
-    cpue_nokm2 = get_opt_float(target['cpue_nokm2'])
-    cpue_no1000km2 = get_opt_float(target['cpue_no1000km2'])
-    weight_kg = get_opt_float(target['weight_kg'])
-    count = get_opt_float(target['count'])
-    bottom_temperature_c = get_opt_float(target['bottom_temperature_c'])
-    surface_temperature_c = get_opt_float(target['surface_temperature_c'])
-    depth_m = float(target['depth_m'])
-    distance_fished_km = float(target['distance_fished_km'])
-    net_width_m = float(target['net_width_m'])
-    net_height_m = float(target['net_height_m'])
-    area_swept_ha = float(target['area_swept_ha'])
-    duration_hr = float(target['duration_hr'])
-    tsn = get_opt_int(target['tsn'])
-    ak_survey_id = int(target['ak_survey_id'])
-
-    return Record(
-        year,
-        srvy,
-        survey,
-        survey_id,
-        cruise,
-        haul,
-        stratum,
-        station,
-        vessel_name,
-        vessel_id,
-        date_time,
-        latitude_dd,
-        longitude_dd,
-        species_code,
-        common_name,
-        scientific_name,
-        taxon_confidence,
-        cpue_kgha,
-        cpue_kgkm2,
-        cpue_kg1000km2,
-        cpue_noha,
-        cpue_nokm2,
-        cpue_no1000km2,
-        weight_kg,
-        count,
-        bottom_temperature_c,
-        surface_temperature_c,
-        depth_m,
-        distance_fished_km,
-        net_width_m,
-        net_height_m,
-        area_swept_ha,
-        duration_hr,
-        tsn,
-        ak_survey_id
-    )
-
-
-class ParseRecordResult:
-    """Object with the results of trying to parse a record from the API.
-
-    Object with the results of trying to parse a record from the API, allowing
-    for internal record keeping within the afscgap library. Note that this is
-    an internal data structure and not expected to reach client code.
-    """
-
-    def __init__(self, raw_record: dict, parsed: typing.Optional[Record]):
-        """Create a new record of a parse attempt.
-
-        Args:
-            raw_record: Item from the API's JSON response payload that the
-                library attempted to parse.
-            parsed: The Record read if successful or None if it could not be
-                parsed.
-        """
-        self._raw_record = raw_record
-        self._parsed = parsed
-
-    def get_raw_record(self) -> dict:
-        """Get the input raw JSON record.
-
-        Returns:
-            Item from the API's JSON response payload that the library attempted
-            to parse.
-        """
-        return self._raw_record
-
-    def get_parsed(self) -> typing.Optional[Record]:
-        """Get the record that was parsed if successful.
-
-        Returns:
-            The Record read if successful or None if it could not be parsed.
-        """
-        return self._parsed
-
-    def meets_requirements(self, allow_incomplete: bool) -> bool:
-        """Determine if this record is "valid" according to the client code.
-
-        Args:
-            allow_incomplete: Flag indicating if incomplete records are
-                considered valid. If true, incomplete records will be considered
-                valid. If false, incomplete records will be considered invalid.
-                Incomplete means missing any optional fields or failing to
-                achieve an ISO 8601 date_time value.
-
-        Returns:
-            True if this record was parsed successfully and meets the
-            requirements specified for being considered "valid" per the
-            allow_incomplete flag. False otherwise.
-        """
-        if self._parsed is None:
-            return False
-
-        return allow_incomplete or self._parsed.is_complete()
-
-
-def try_parse(target: dict) -> ParseRecordResult:
-    """Attempt parsing a Record from an input item dictionary from the API.
-
-    Params:
-        target: The dictionary from which to parse. This should be an item
-            from the items array in the returned JSON payload from the API.
-
-    Returns:
-        Parse result describing if the dictionary was parsed successfully.
-    """
-    try:
-        return ParseRecordResult(target, parse_record(target))
-    except (ValueError, KeyError):
-        return ParseRecordResult(target, None)
