+++ afscgap/test/test_client.py
@@ -1,8 +1,8 @@
 """
 Unit tests for the logic that interacts with the API and makes HTTP requests.
 
-(c) 2023 The Eric and Wendy Schmidt Center for Data Science and the Environment
-at UC Berkeley.
+(c) 2023 Regents of University of California / The Eric and Wendy Schmidt Center
+for Data Science and the Environment at UC Berkeley.
 
 This file is part of afscgap released under the BSD 3-Clause License. See
 LICENSE.txt.
